36teen upload

+++ app/src/main/java/com/bsel/remitngo/bottomSheet/SaveRecipientBottomSheet.kt
@@ -199,7 +199,7 @@ class SaveRecipientBottomSheet : BottomSheetDialogFragment() {
 
         val beneficiaryItem = BeneficiaryItem(
             deviceId = deviceId,
-            personId = personId.toInt(),
+            personId = personId,
             firstname = beneAccountName,
             middlename = "",
             lastname = "",
